Fix | deleted test view
